requirements.txt
pandas==0.24.2
numpy==1.16.4
torchvision==0.2.1
scipy==1.4.1
seaborn==0.9.0
torch==1.5.1
matplotlib==2.2.2
Pillow==7.2.0
pytest==3.5.1
tensorboard==2.2.2